Adds containerization process for the application
Adds the worker and client Dockerfiles.

Adds the all-build-images make target, which which will build all the container
images for all the OS/ARCH combination without pushing the images.

Adds the all-publish-images make target, which which will build and publish all
the container images for all the OS/ARCH combination and pushes the images.

Adds the push-manifests make target, which depends on the all-publish-images target.
In addition to that, this target will create the manifest lists for the application images.